requirements.txt
discord.py
wand
requests
Pillow
git+https://github.com/vasilysinitsin/Faces.git